+++ src/lib/auth.ts
@@ -34,30 +34,3 @@ export const authLogout = () => {
   cookies().delete("auth");
   redirect("/");
 };
-
-export const auth = async (cookie: string | undefined) => {
-  const token = cookie;
-  if (!token) {
-    return false;
-  }
-  try {
-    const decoded = await jwtVerify(
-      token,
-      new TextEncoder().encode(getJwtSecretKey()),
-    );
-    console.log("decoded ===", decoded);
-    if (!decoded || !decoded.payload.exp) {
-      throw new Error("JWT token is not valid");
-    }
-    if (decoded.payload.exp > Math.floor((Date.now() / 1000) * 1000 * 60)) {
-      return false;
-    }
-  } catch (err) {
-    return false;
-  }
-  // const authToken
-  // check for auth cookie
-  return true;
-};
-
-export const generateToken = () => {};
